+++ src/main/java/com/yanny/ytech/compatibility/TopCompatibility.java
@@ -62,7 +62,7 @@ public class TopCompatibility {
                             addPrimitiveSmelterInfo(probeInfo, blockEntity);
                         } else if (entity instanceof IIrrigationBlockEntity blockEntity) {
                             addIrrigationInfo(probeInfo, blockEntity);
-                        } else if (entity instanceof PottersWheelBlockEntity blockEntity) {
+                        } else if (entity instanceof PottersWheelBlockEntity blockEntity && !blockEntity.getItem().isEmpty()) {
                             probeInfo.horizontal().item(blockEntity.getItem());
                         } else if (entity instanceof FirePitBlockEntity blockEntity && !blockEntity.getItem().isEmpty()) {
                             probeInfo.horizontal().text(Component.translatable("text.ytech.top.fire_pit.progress", Integer.toString(blockEntity.getProgress())));
@@ -71,6 +71,12 @@ public class TopCompatibility {
                         } else if (entity instanceof AmphoraBlockEntity blockEntity) {
                             probeInfo.horizontal().item(blockEntity.getItem());
                         } else if (entity instanceof WoodenBoxBlockEntity blockEntity) {
+                            ItemStack item = blockEntity.getItem(probeHitData.getPos(), probeHitData.getSideHit(), probeHitData.getHitVec());
+
+                            if (!item.isEmpty()) {
+                                probeInfo.horizontal().item(item).text(Component.translatable(item.getDescriptionId()));
+                            }
+                        } else if (entity instanceof ToolRackBlockEntity blockEntity) {
                             ItemStack item = blockEntity.getItem(probeHitData.getPos(), probeHitData.getSideHit(), probeHitData.getHitVec());
 
                             if (!item.isEmpty()) {