+++ src/main/java/com/yanny/ytech/generation/YTechItemTagsProvider.java
@@ -89,6 +89,7 @@ class YTechItemTagsProvider extends ItemTagsProvider {
         tag(YTechItemTags.THATCH).add(YTechItems.THATCH.get());
         tag(YTechItemTags.THATCH_SLABS).add(YTechItems.THATCH_SLAB.get());
         tag(YTechItemTags.THATCH_STAIRS).add(YTechItems.THATCH_STAIRS.get());
+        tag(YTechItemTags.TOOL_RACKS).add(YTechItems.TOOL_RACK.get());
         tag(YTechItemTags.TREE_STUMPS).add(YTechItems.TREE_STUMP.get());
         tag(YTechItemTags.WOODEN_BOXES).add(YTechItems.WOODEN_BOX.get());